refactored particle system code, modified the scaling function for the particle size
